[1131] Extract Withdrawing into its own service and set correct DTTP fields
